rename `mclient.py` -> `cypher.py`
