+++ megadl/helpers/cypher.py
